Migrations adjustments
