+++ sources/packages/api/src-sql/MSFAANumbers/Create-msfaa-numbers-table.sql
