+++ sources/packages/api/src/database/entities/index.ts
@@ -20,4 +20,4 @@ export * from "./assessment-status.type";
 export * from "./coe-status.type";
 export * from "./program-year.model";
 export * from "./pir-denied-reason.model";
-export * from "./msfaa-number";
+export * from "./msfaa-number.model";
